# AsynCUDA13.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AsynCUDA13.API v1",
    "version": "v1"
  },
  "paths": {
    "/execute-image/{kernelName}/{imageIdOrNameOrPath}": {
      "post": {
        "tags": [
          "ImageKernel"
        ],
        "parameters": [
          {
            "name": "imageIdOrNameOrPath",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kernelName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "argumentValues",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "overwriteImage",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "unloadKernelAfterExecution",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageData"
                }
              }
            }
          }
        }
      }
    },
    "/execute-image-file": {
      "post": {
        "tags": [
          "ImageKernel"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "kernelName": {
                    "type": "string"
                  },
                  "argumentValues": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "overwriteImage": {
                    "type": "boolean",
                    "default": true
                  },
                  "unloadKernelAfterExecution": {
                    "type": "boolean",
                    "default": false
                  }
                }
              },
              "encoding": {
                "imageFile": {
                  "style": "form"
                },
                "kernelName": {
                  "style": "form"
                },
                "argumentValues": {
                  "style": "form"
                },
                "overwriteImage": {
                  "style": "form"
                },
                "unloadKernelAfterExecution": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/execute-image-scoped/edge_detection": {
      "post": {
        "tags": [
          "ImageKernel"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "thickness": {
                    "type": "integer",
                    "format": "int32",
                    "default": 1
                  },
                  "threshold": {
                    "type": "number",
                    "format": "float",
                    "default": 0.125
                  },
                  "edgeR": {
                    "type": "integer",
                    "format": "int32",
                    "default": 255
                  },
                  "edgeG": {
                    "type": "integer",
                    "format": "int32",
                    "default": 0
                  },
                  "edgeB": {
                    "type": "integer",
                    "format": "int32",
                    "default": 0
                  },
                  "deviceId": {
                    "type": "integer",
                    "format": "int32",
                    "default": 0
                  },
                  "kernelVersion": {
                    "type": "string",
                    "default": ""
                  }
                }
              },
              "encoding": {
                "imageFile": {
                  "style": "form"
                },
                "thickness": {
                  "style": "form"
                },
                "threshold": {
                  "style": "form"
                },
                "edgeR": {
                  "style": "form"
                },
                "edgeG": {
                  "style": "form"
                },
                "edgeB": {
                  "style": "form"
                },
                "deviceId": {
                  "style": "form"
                },
                "kernelVersion": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImageData": {
        "required": [
          "info"
        ],
        "type": "object",
        "properties": {
          "info": {
            "$ref": "#/components/schemas/ImageInfo"
          },
          "pointer": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          },
          "base64Data": {
            "type": "string",
            "nullable": true
          },
          "base64Image": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "dataSizeMb": {
            "type": "number",
            "format": "float",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ImageInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "mediaType": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "channels": {
            "type": "integer",
            "format": "int32"
          },
          "originalSizeMb": {
            "type": "number",
            "format": "float"
          },
          "pointer": {
            "type": "string",
            "nullable": true
          },
          "onGpu": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}